+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is used to list changes made in each version of the mac-app-store cook
 - Remove all references to reattach-to-user-namespace
 - Switch to a custom resource-only, recipe-less model
 - Drop support for installing Mas outside of Homebrew
+- Remove the Mas sign_in and sign_out actions
 
 ## 3.1.0 (2017-04-20)